Refactor to replace row/column with newtypes
This is a pretty major refactor, following the newtype pattern from the
advanced rust chapter of the rust book. I'm not sure it's actually a
worthwhile improvement just yet, there is a bunch of additional
complexity and I've had to use deref in a few places which goes against
the point of this a bit. I'll see how I feel about it in time.