Implement negative heuristic for self checked
Also had to remove the kings from the gain 9 material better than
gaining 5 test, because it was convinced that attacking the king was the
right thing to do. It is difficult to show it that that is not the best
move without going into a deeper search, which I don't want to do yet.