Refactor heuristics: always same value
I completely misarchitected the heuristics, assuming that it was
significant which player wanted to do the evaluation (i.e. which one was
the AI). Once I started learning more about search I realised this was
completely wrong and I'd written a load of unnecessary code.

Some tests that would likely be broken by this refactor are commented
out because search has not yet been modified to account for it.